feat: add Prometheus hooks and error reporting to odom providers
- Add _on_odom_processed() and _on_odom_error() hooks to OdomProviderBase
- Override hooks in all odom subclasses for Prometheus heartbeat/error reporting
- Add Prometheus assertion tests to provider test files
- Merge integration tests into test_prometheus_monitor.py
