+++ src/providers/tron_odom_provider.py
@@ -166,3 +166,10 @@ class TronOdomProvider(OdomProviderBase):
         Sends a heartbeat to the Prometheus monitor.
         """
         self._monitor.heartbeat("TronOdomProvider")
+
+    def _on_odom_error(self, error: str):
+        """
+        Hook called when an error occurs during odometry processing.
+        Reports the error to the Prometheus monitor.
+        """
+        self._monitor.report_error("TronOdomProvider", error)
